+++ harness/tests/src/main/kotlin/godot/tests/Invocation.kt
@@ -40,6 +40,7 @@ import godot.core.Vector2
 import godot.core.Vector3
 import godot.core.dictionaryOf
 import godot.core.variantArrayOf
+import godot.extensions.node
 import godot.signals.signal
 import godot.tests.subpackage.OtherScript
 import godot.util.RealT
@@ -364,10 +365,16 @@ class Invocation : Spatial() {
 		name = formerName
 		val test = DateTime.now() //external dependency to test dependency inclusion in dummyCompilation
 
+        val getNode: Button? = node("CanvasLayer/Button")
+
+        val path = NodePath("CanvasLayer/Button")
+        val getNode2: Button? = node(path)
+
 		signalNoParam.connect(invocation, invocation::hookNoParam)
 		signalOneParam.connect(invocation, invocation::hookOneParam)
 		signalTwoParam.connect(invocation, invocation::hookTwoParam)
-		(getNodeOrNull(NodePath("CanvasLayer/Button")) as Button?)?.signalPressed?.connect(
+
+		(getNodeOrNull(path) as Button?)?.signalPressed?.connect(
 			invocation,
 			invocation::hookNoParam
 		)
